chore: prepare GUID-linked preview fix
