+++ Tests/IMsgCoreTests/MessageStoreURLPreviewTests.swift
@@ -4,77 +4,6 @@ import Testing
 
 @testable import IMsgCore
 
-func makeURLPreviewTestDB() throws -> Connection {
-  let db = try Connection(.inMemory)
-  try db.execute(
-    """
-    CREATE TABLE message (
-      ROWID INTEGER PRIMARY KEY,
-      handle_id INTEGER,
-      text TEXT,
-      guid TEXT,
-      associated_message_guid TEXT,
-      associated_message_type INTEGER,
-      reply_to_guid TEXT,
-      balloon_bundle_id TEXT,
-      is_read INTEGER, date_read INTEGER,
-      date INTEGER,
-      is_from_me INTEGER,
-      service TEXT
-    );
-    """
-  )
-  try db.execute("CREATE TABLE handle (ROWID INTEGER PRIMARY KEY, id TEXT);")
-  try db.execute("CREATE TABLE chat_message_join (chat_id INTEGER, message_id INTEGER);")
-  try db.execute(
-    "CREATE TABLE message_attachment_join (message_id INTEGER, attachment_id INTEGER);")
-  return db
-}
-
-func insertURLPreviewTestMessage(
-  _ db: Connection,
-  rowID: Int64,
-  chatID: Int64 = 1,
-  handleID: Int64 = 1,
-  text: String,
-  guid: String,
-  associatedMessageGUID: String? = nil,
-  associatedMessageType: Int? = nil,
-  replyToGUID: String? = nil,
-  balloonBundleID: String? = nil,
-  date: Date,
-  isFromMe: Bool = false,
-  isRead: Bool = false,
-  dateRead: Date? = nil
-) throws {
-  try db.run(
-    """
-    INSERT INTO message(
-      ROWID, handle_id, text, guid, associated_message_guid, associated_message_type,
-      reply_to_guid, balloon_bundle_id, is_read, date_read, date, is_from_me, service
-    )
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 'iMessage')
-    """,
-    rowID,
-    handleID,
-    text,
-    guid,
-    associatedMessageGUID,
-    associatedMessageType,
-    replyToGUID,
-    balloonBundleID,
-    isRead ? 1 : 0,
-    dateRead.map(TestDatabase.appleEpoch) ?? 0,
-    TestDatabase.appleEpoch(date),
-    isFromMe ? 1 : 0
-  )
-  try db.run(
-    "INSERT INTO chat_message_join(chat_id, message_id) VALUES (?, ?)",
-    chatID,
-    rowID
-  )
-}
-
 @Test
 func messagesAfterSkipsPreviewOnlyBatchForPublicCursorlessAPI() throws {
   let db = try makeURLPreviewTestDB()
@@ -245,100 +174,6 @@ func messagesAfterCoalescesURLPreviewSplitSend() throws {
     balloonBundleID: MessageStore.urlPreviewBalloonBundleID,
     date: now.addingTimeInterval(1)
   )
-
-  let store = try MessageStore(connection: db, path: ":memory:")
-  let messages = try store.messagesAfter(afterRowID: 0, chatID: 1, limit: 10)
-
-  #expect(messages.map(\.rowID) == [1])
-  #expect(messages.first?.urlPreview?.rowID == 2)
-}
-
-@Test
-func messagesAfterCoalescesGUIDLinkedURLPreviewWhenTextOmitsURL() throws {
-  let db = try makeURLPreviewTestDB()
-  let now = Date()
-  try db.run("INSERT INTO handle(ROWID, id) VALUES (1, '+123')")
-  try insertURLPreviewTestMessage(
-    db,
-    rowID: 1,
-    text: "Check this out",
-    guid: "text-guid",
-    date: now
-  )
-  try insertURLPreviewTestMessage(
-    db,
-    rowID: 2,
-    text: "https://example.com",
-    guid: "preview-guid",
-    replyToGUID: "p:0/text-guid",
-    balloonBundleID: MessageStore.urlPreviewBalloonBundleID,
-    date: now.addingTimeInterval(0.399)
-  )
-
-  let store = try MessageStore(connection: db, path: ":memory:")
-  let messages = try store.messagesAfter(afterRowID: 0, chatID: 1, limit: 10)
-
-  #expect(messages.map(\.rowID) == [1])
-  #expect(messages.first?.text == "Check this out\nhttps://example.com")
-  #expect(messages.first?.urlPreview?.rowID == 2)
-}
-
-@Test
-func messagesAfterKeepsUnlinkedURLPreviewWhenTextOmitsURL() throws {
-  let db = try makeURLPreviewTestDB()
-  let now = Date()
-  try db.run("INSERT INTO handle(ROWID, id) VALUES (1, '+123')")
-  try insertURLPreviewTestMessage(
-    db,
-    rowID: 1,
-    text: "Check this out",
-    guid: "text-guid",
-    date: now
-  )
-  try insertURLPreviewTestMessage(
-    db,
-    rowID: 2,
-    text: "https://example.com",
-    guid: "preview-guid",
-    balloonBundleID: MessageStore.urlPreviewBalloonBundleID,
-    date: now.addingTimeInterval(0.206)
-  )
-
-  let store = try MessageStore(connection: db, path: ":memory:")
-  let messages = try store.messagesAfter(afterRowID: 0, chatID: 1, limit: 10)
-
-  #expect(messages.map(\.rowID) == [1, 2])
-  #expect(messages.allSatisfy { $0.urlPreview == nil })
-}
-
-@Test
-func messagesAfterCoalescesGUIDLinkedPreviewWithoutReactionColumns() throws {
-  let db = try Connection(.inMemory)
-  try MessageDatabaseFixture.createSchema(
-    db,
-    options: MessageDatabaseFixture.SchemaOptions(
-      includeGUID: true,
-      includeBalloonBundleID: true,
-      includeReplyToGUID: true
-    )
-  )
-  let now = Date()
-  try db.run("INSERT INTO handle(ROWID, id) VALUES (1, '+123')")
-  try db.run(
-    """
-    INSERT INTO message(
-      ROWID, handle_id, text, guid, reply_to_guid, balloon_bundle_id,
-      date, is_from_me, service
-    )
-    VALUES
-      (1, 1, 'Check this out', 'text-guid', NULL, NULL, ?, 0, 'iMessage'),
-      (2, 1, 'https://example.com', 'preview-guid', 'text-guid', ?, ?, 0, 'iMessage')
-    """,
-    TestDatabase.appleEpoch(now),
-    MessageStore.urlPreviewBalloonBundleID,
-    TestDatabase.appleEpoch(now.addingTimeInterval(0.399))
-  )
-  try db.run("INSERT INTO chat_message_join(chat_id, message_id) VALUES (1, 1), (1, 2)")
 
   let store = try MessageStore(connection: db, path: ":memory:")
   let messages = try store.messagesAfter(afterRowID: 0, chatID: 1, limit: 10)
